feat: add unittests

+++ DiveBlenderApi/GasMixture.cs
@@ -2,10 +2,10 @@ namespace DiveBlenderApi
 {
     public class GasMixture {
         public int Pressure {get; set;}
-        public float Oxygen {get; set; }
-        public float Helium {get; set;}
+        public double Oxygen {get; set; }
+        public double Helium {get; set;}
 
-        public GasMixture(int pressure, float oxygen, float helium)
+        public GasMixture(int pressure, double oxygen, double helium)
         {
             Pressure = pressure; 
             Oxygen = oxygen;
